pyproject.toml
[project]
name = "lesson-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi>=0.115.8",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
]
